+++ deprecated-react-native-prop-types/CHANGELOG.md
@@ -6,3 +6,4 @@
   - Includes all of `ViewPropTypes`.
   - Complete checks for `defaultSource` and `source`.
   - Added `alt`, `crossOrigin`, `height`, `referrerPolicy`, `src`, `srcSet`, `tintColor`, and `width`.
+  - Added `objectFit` to `style`.